Updated the bootstrap, etcd, loadbalancer poller utilization
As HAProxy has no github release so made it to use 3.0-LTS
